Silent check warning on DT import
All the code for the Shiny app is in `inst` and so is invisible to the
check (it's not technically package code). DT is not used anywhere else
nor in explicit namespace imports, which leads to the warning. Added a
dummy function that contains a call to a function from DT so that the
check sees that DT is used at all.

References #103.